refactor(recall): relocate stateless recall-memory-map out of orchestration
recall-memory-map.ts is a pure identity/dedup helper with no session or lifecycle state, but living under orchestration/ pulled it into the lifecycle-matrix coverage glob and failed the gate (it has no LifecycleSubject because it has no lifecycle behavior). Move it to packages/remnic-core/src/ alongside the other stateless helpers and update the two importers.
